#17 Completed
The Document Writer Generates Bitmaps which are stored in a Dictionary with String, Bitmap keyvalue pairs, where the string represents the name of the graph, and the bitmap being the graph. When writing the HTML, the Writer creates all the bitmap are saved into PNGs in the {outputPath}/objGraphs/{name}_Graph.png, and are added to the page dynamically, on whether those bitmaps exist in the dictionary.